requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect==1.3.0
distro==1.9.0
fastapi==0.115.13
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
firecrawl-py==2.8.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.10.0
joblib==1.5.1
lxml==5.4.0
lxml_html_clean==0.4.2
multidict==6.5.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
openai==1.88.0
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
yarl==1.20.1
